requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-filter>=23.3
channels>=4.0.0
channels-redis>=4.1.0
redis>=5.0.1
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-django>=4.5.2
coverage>=7.4.0
requests>=2.31.0